# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  http.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: integer
    type: object
...
  models.Order:
    properties:
      created_at:
        type: string
      id:
        type: string
      payment_method:
        type: string
      shipping_address:
        $ref: '#/definitions/models.ShippingAddress'
      status:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      order_id:
        type: string
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  models.Product:
    properties:
      category_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  models.ShippingAddress:
    properties:
      city:
        type: string
      country:
        type: string
      postalCode:
        type: string
      street:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      password_hash:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  order.CreateOrderRequset:
    properties:
      items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      payment_method:
        enum:
        - card
        - cash
        type: string
      shipping_adress:
        $ref: '#/definitions/models.ShippingAddress'
      user_id:
        type: string
    required:
    - items
    - payment_method
    - shipping_adress
    - user_id
    type: object
  order.OrderListResponse:
    properties:
      order:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      total:
        type: integer
    type: object
  order.UpdateOrderRequest:
    properties:
      status:
        enum:
        - pending
        - paid
        - shipped
        - canceled
        - delivered
        type: string
    required:
    - status
    type: object
  product.CreateProductRequest:
    properties:
      category_id:
        type: string
      description:
        maxLength: 1000
        type: string
      image_url:
        type: string
      name:
        maxLength: 255
        minLength: 3
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
    required:
    - name
    - price
    - stock
    type: object
  product.ProductListResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      total:
        type: integer
    type: object
  product.UpdateProductRequest:
    properties:
      category_id:
        type: string
      description:
        maxLength: 1000
        type: string
      image_url:
        type: string
      name:
        maxLength: 255
        minLength: 3
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
    type: object
  user.UpdateUserRequest:
    properties:
      email:
        type: string
      first_name:
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
      role:
        enum:
        - user
        - admin
        type: string
    type: object
  user.UserListResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для e-commerce платформи з управлінням продуктами, кошиком та замовленнями
  title: E-Commerce API
  version: "1.0"
paths:
  /admin/orders:
    get:
      consumes:
      - application/json
      description: Повертає список всіх замовлень з фільтром по статусу та пагінацією
      parameters:
      - description: Статус замовлення (pending, paid, shipped, canceled, delivered)
        in: query
        name: status
        type: string
      - default: 20
        description: Кількість елементів на сторінку
        in: query
        name: limit
        type: integer
      - default: 0
        description: Зміщення для пагінації
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.OrderListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отримати всі замовлення (Admin)
      tags:
      - admin
  /admin/orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Змінює статус конкретного замовлення
      parameters:
      - description: ID замовлення
        in: path
        name: id
        required: true
        type: string
      - description: Новий статус замовлення
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/order.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Invalid request body or ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Оновлення статусу замовлення (Admin)
      tags:
      - admin
  /admin/products:
    post:
      consumes:
      - application/json
      description: Створює новий продукт
      parameters:
      - description: Дані продукту
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/product.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Створення продукту (Admin)
      tags:
      - admin
  /admin/products/{id}:
    put:
      consumes:
      - application/json
      description: Оновлює існуючий продукт
      parameters:
      - description: ID продукту
        in: path
        name: id
        required: true
        type: string
      - description: Дані для оновлення
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/product.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid request body or ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Оновлення продукту (Admin)
      tags:
      - admin
  /admin/users:
    get:
      consumes:
      - application/json
      description: Повертає список користувачів з фільтром та пагінацією
      parameters:
      - description: Пошуковий запит по імені або email
        in: query
        name: search
        type: string
      - description: Роль користувача (user, admin)
        in: query
        name: role
        type: string
      - default: 20
        description: Кількість елементів на сторінку
        in: query
        name: limit
        type: integer
      - default: 0
        description: Зміщення для пагінації
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отримати список користувачів (Admin)
      tags:
      - admin
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: Видаляє конкретного користувача
      parameters:
      - description: ID користувача
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Видалення користувача (Admin)
      tags:
      - admin
    get:
      consumes:
      - application/json
      description: Повертає дані конкретного користувача
      parameters:
      - description: ID користувача
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отримати користувача за ID (Admin)
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Оновлює дані конкретного користувача
      parameters:
      - description: ID користувача
        in: path
        name: id
        required: true
        type: string
      - description: Дані для оновлення
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid request body or ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Оновлення користувача (Admin)
      tags:
      - admin
...
  /orders:
    get:
      consumes:
      - application/json
      description: Повертає список замовлень користувача з можливістю фільтрації за
        статусом та пагінацією
      parameters:
      - description: Фільтр по статусу
        enum:
        - pending
        - paid
        - shipped
        - canceled
        - delivered
        in: query
        name: status
        type: string
      - default: 20
        description: Кількість елементів на сторінку
        in: query
        name: limit
        type: integer
      - default: 0
        description: Зміщення для пагінації
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.OrderListResponse'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: User not authorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отримати список замовлень
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Створює нове замовлення для авторизованого користувача
      parameters:
      - description: Дані замовлення
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order.CreateOrderRequset'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: User not authorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Створити замовлення
      tags:
      - orders
...
  /products:
    get:
      consumes:
      - application/json
      description: Повертає список продуктів з можливістю фільтрації за категорією,
        ціною, наявністю та пагінацією
      parameters:
      - description: Category ID (UUID)
        in: query
        name: category_id
        type: string
      - description: Мінімальна ціна
        in: query
        name: min_price
        type: number
      - description: Максимальна ціна
        in: query
        name: max_price
        type: number
      - description: Пошуковий запит
        in: query
        name: search
        type: string
      - description: Тільки товари в наявності
        in: query
        name: in_stock
        type: boolean
      - default: 20
        description: Кількість елементів на сторінку
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Зміщення для пагінації
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.ProductListResponse'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Отримати список продуктів
      tags:
      - products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: Повертає детальну інформацію про продукт за його унікальним ідентифікатором
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Отримати продукт за ID
      tags:
      - products
  /products/search:
    get:
      consumes:
      - application/json
      description: Повертає список продуктів, що відповідають пошуковому запиту
      parameters:
      - description: Пошуковий запит
        in: query
        name: q
        required: true
        type: string
      - default: 20
        description: Кількість елементів на сторінку
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Зміщення для пагінації
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "400":
          description: Search query is required or invalid parameters
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Пошук продуктів
      tags:
      - products
  /users:
    get:
      consumes:
      - application/json
      description: Повертає дані користувача, який авторизований
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: User not authorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отримати інформацію про користувача
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Оновлює дані користувача (email, пароль, ім'я, прізвище, роль)
      parameters:
      - description: Дані для оновлення користувача
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid request body or invalid fields
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: User not authorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Оновити дані користувача
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"